aider: Fix the issue by specifying unique binary directories for each `add_subdirectory` call in the root `CMakeLists.txt`.
# Aider chat conversation:

USER:  For the root `CMakeLists.txt` when I try to build in CLion I get error
CMake Error at CMakeLists.txt:14 (add_subdirectory):
  The binary directory

    /Users/asimi/workplace/understanding-software-dynamics/cmake-build-debug/chapters

  is already used to build a source directory.  It cannot be used to build
  source directory

    /Users/asimi/workplace/understanding-software-dynamics/chapters

  Specify a unique binary directory name.

ASSISTANT: To resolve the issue, we need to ensure that each `add_subdirectory` call in the root `CMakeLists.txt` specifies a unique binary directory. We can achieve this by using the `BINARY_DIR` option in the `add_subdirectory` command.

Here is the diff to fix the issue:

```diff
--- CMakeLists.txt
+++ CMakeLists.txt
@@ ... @@
-add_subdirectory(chapters)
+add_subdirectory(chapters ${CMAKE_BINARY_DIR}/chapters)
```