aider: Fix build directory location and make command execution in `scripts/build-mac-aarch64.sh` and `CMakeLists.txt`.
# Aider chat conversation:

USER:  It is building the makefile but it's putting it in the root directory, I think the wrong place?
$ cmake -DCMAKE_TOOLCHAIN_FILE=../../cmake/toolchains/aarch64.cmake ../..
-- Configuring done (0.0s)
-- Generating done (0.0s)
-- Build files have been written to: /Users/asimi/workplace/understanding-software-dynamics

ASSISTANT: It looks like the `cmake` command is not correctly setting the build directory. We need to ensure that the `cmake` command is executed from within the `build/aarch64` directory and that it correctly references the source directory.

Let's update the `scripts/build-mac-aarch64.sh` script to fix this issue:

```diff
--- scripts/build-mac-aarch64.sh
+++ scripts/build-mac-aarch64.sh
@@ ... @@
-cmake -DCMAKE_TOOLCHAIN_FILE=../../cmake/toolchains/aarch64.cmake ../..
+cmake -DCMAKE_TOOLCHAIN_FILE=../../cmake/toolchains/aarch64.cmake ../..
+make
```

Additionally, we should ensure that the `CMakeLists.txt` file in the root directory correctly sets the build directory:

```diff
--- CMakeLists.txt
+++ CMakeLists.txt
@@ ... @@
+set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
```

This change ensures that the build files are generated in the correct directory and that the `make` command is executed properly.